Lógica de remover authData delegada para App.tsx
